\name{LDdiag}
\alias{LDdiag-package}
\docType{package}
\title{Link Function and Distribution Diagnostic Test for Social Science Researchers}

\description{
Social science researchers sometimes have difficulties in deciding appropriate regression methods and/or 
link functions for their studies. This package provides some model diagonistic approaches for 
link function and distribution, such as Pregibon linearity test, Hosmer-Lemeshow test, Copas test, Park test, etc.
}

\details{
\tabular{ll}{
Package: \tab LDdiag\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-05-11\cr
License: \tab GPL>=2\cr
LazyLoad: \tab yes\cr
}
}

\author{ Yongmei Ni <yongmei.ni@utah.edu>  }

\keyword{package}

